- create basic my app with ScreenUtilInit - add package helpers and create spacing file
