+++ lib/core/helpers/spacing.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+SizedBox verticalSpace(double height) => SizedBox(height: height.h);
+
+SizedBox horizontalSpace(double width) => SizedBox(width: width.w);
